requirements.txt
asgiref==3.7.2
boto3==1.35.95
botocore==1.35.95
Django==4.2.16
django-jazzmin==3.0.1
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.4.4
urllib3<1.27,>=1.25.4
whitenoise==6.8.2
